Launches an Azure instance and regist a DNS record with its public IP. (#718)
* Move gcp and aws dns operations to a dedicated file.
